+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
-TODO: You can follow this format for your changelog: <https://keepachangelog.com/en/1.1.0/>
+## 0.1.0
+
+- Added Banana Peel Join Fix
+- Added Maximum Voice Volume Tweak
+- Added Campfire Hunger Reduction